Enable symlinks on msys in configure
MSYS's default behavior when trying to create a symlink is to copy the
source file/folder to the destination since there's no reliable way to
create a symlink visible to Windows-native (non-cygwin) programs.

The configure script tries to symlink lib64 to lib, which fails with
that default behavior, so change the settings so MSYS generates
Cygwin-only symlinks so that we can make lib64 be a symlink.

It remains to be seen whether enabling symlinks will cause problems at a
later point due to the fact that they won't be symlinks on Windows.