Avoid different extension names for text Python extension
When building a small C extension to check if distutils is correctly
installed, change the extension name in the setup Python script to match
the name for the module specified in the C/C++ file.

Python on Linux doesn't seem to have a problem with those 2 names being
different, but having different names there on Windows results in an
error, so fix that by just using the same name in both places.